put the texture cache into it's own file
