+++ src/GPU.h
@@ -617,6 +617,4 @@ void SetDispStat(u32 cpu, u16 val);
 void SetVCount(u16 val);
 }
 
-#include "GPU3D.h"
-
 #endif
